+++ internal/server/server_functions.go
@@ -3,168 +3,20 @@ package server
 import (
 	"context"
 	"encoding/json"
-	"fmt"
-	"html/template"
-	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"os"
-
 	"strings"
-	"sync"
 	"time"
 
-	"cod/cluster"
-	"cod/events"
-	"cod/logger"
-	"cod/logs"
-	"cod/utils"
+	"cod/internal/cluster"
+	"cod/internal/events"
+	"cod/internal/logger"
+	"cod/internal/logs"
 
-	"github.com/gorilla/websocket"
+	"net/http"
+
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
 )
-
-type Client struct {
-	conn            *websocket.Conn
-	watchEventslist map[string]bool
-	watchLogs       bool
-	logSessionId    string
-	eventSessionId  string
-}
-
-type Server struct {
-	clusters           []string
-	clusterConditions  map[string][]map[string]interface{}
-	upgrader           websocket.Upgrader
-	clients            map[*Client]bool
-	eventWatchers      map[string]context.CancelFunc
-	logWatcher         context.CancelFunc
-	broadcast          chan utils.Message
-	clientsMutex       sync.Mutex
-	logMutex           sync.Mutex
-	eventWatchersMutex sync.Mutex
-	eventCache         map[string][]utils.Message // Map of clusterName to cached events
-	eventCacheMutex    sync.RWMutex
-	clientset          *kubernetes.Clientset
-	dynamicClient      dynamic.Interface
-}
-
-func NewServer() *Server {
-	return &Server{
-		upgrader:          websocket.Upgrader{ReadBufferSize: 1024, WriteBufferSize: 1024},
-		clients:           make(map[*Client]bool),
-		eventWatchers:     make(map[string]context.CancelFunc),
-		broadcast:         make(chan utils.Message),
-		clusterConditions: make(map[string][]map[string]interface{}),
-		eventCache:        make(map[string][]utils.Message),
-	}
-}
-
-func (s *Server) Start() {
-	var err error
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		logger.Log.Fatal("Failed to get in-cluster config", zap.Error(err))
-	}
-
-	s.clientset, err = kubernetes.NewForConfig(config)
-	if err != nil {
-		logger.Log.Fatal("Failed to create Kubernetes client", zap.Error(err))
-	}
-
-	s.dynamicClient, err = dynamic.NewForConfig(config)
-	if err != nil {
-		logger.Log.Fatal("Failed to create dynamic client", zap.Error(err))
-	}
-
-	namespace := os.Getenv("WATCH_NAMESPACE")
-	if namespace == "" {
-		logger.Log.Fatal("WATCH_NAMESPACE environment variable not set")
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	// Start the cluster watcher with separate add and delete functions
-	go cluster.StartClusterWatcher(ctx, s.dynamicClient, s.addCluster, s.deleteCluster, s.updateConditions)
-
-	fs := http.FileServer(http.Dir("static"))
-	http.Handle("/static/", http.StripPrefix("/static/", fs))
-
-	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		// First check if this is an API request
-		if s.isAPIRequest(r) {
-			s.handleCouchbaseAPIProxy(w, r)
-			return
-		}
-
-		// Otherwise serve the dashboard
-		tmpl, _ := template.ParseFiles("templates/index.html")
-		tmpl.Execute(w, s.clusters)
-	})
-
-	// Add a handler for cluster-specific pages
-	http.HandleFunc("/cluster/", func(w http.ResponseWriter, r *http.Request) {
-		// Extract the cluster name from the URL path
-		pathParts := strings.Split(r.URL.Path, "/")
-		if len(pathParts) < 3 {
-			http.NotFound(w, r)
-			return
-		}
-
-		clusterName := pathParts[2]
-
-		// Check if the cluster exists in our tracked clusters list
-		clusterExists := false
-		for _, name := range s.clusters {
-			if name == clusterName {
-				clusterExists = true
-				break
-			}
-		}
-
-		if !clusterExists {
-			http.NotFound(w, r)
-			return
-		}
-
-		// Render the cluster template with the cluster name
-		tmpl, err := template.ParseFiles("templates/cluster.html")
-		if err != nil {
-			logger.Log.Error("Error parsing cluster template", zap.Error(err))
-			http.Error(w, "Internal server error", http.StatusInternalServerError)
-			return
-		}
-
-		data := struct {
-			Name string
-		}{
-			Name: clusterName,
-		}
-
-		if err := tmpl.Execute(w, data); err != nil {
-			logger.Log.Error("Error executing cluster template", zap.Error(err))
-			http.Error(w, "Internal server error", http.StatusInternalServerError)
-			return
-		}
-	})
-
-	http.HandleFunc("/ws", s.handleConnections)
-
-	// Add a handler for the Couchbase UI proxy - maintain trailing slash to ensure path consistency
-	http.HandleFunc("/cui/", s.handleCouchbaseUIProxy)
-
-	go s.handleMessages()
-
-	logger.Log.Info("Server started on :3000")
-	if err := http.ListenAndServe(":3000", nil); err != nil {
-		logger.Log.Fatal("Server failed", zap.Error(err))
-	}
-}
 
 // isAPIRequest determines if the request is for a Couchbase API endpoint
 func (s *Server) isAPIRequest(r *http.Request) bool {
@@ -214,81 +66,6 @@ func (s *Server) isAPIRequest(r *http.Request) bool {
 	}
 
 	return false
-}
-
-// handleCouchbaseAPIProxy handles API requests that need to be forwarded to the active Couchbase cluster
-func (s *Server) handleCouchbaseAPIProxy(w http.ResponseWriter, r *http.Request) {
-	// Determine which cluster to use by checking the referer header
-	referer := r.Header.Get("Referer")
-	if referer == "" {
-		http.Error(w, "API requests require a Referer header to determine target cluster", http.StatusBadRequest)
-		return
-	}
-
-	// Extract cluster name from the referer
-	// Expected format: http://host:port/cui/clustername/...
-	refererURL, err := url.Parse(referer)
-	if err != nil {
-		http.Error(w, "Invalid referer URL", http.StatusBadRequest)
-		return
-	}
-
-	refPath := refererURL.Path
-	if !strings.HasPrefix(refPath, "/cui/") {
-		http.Error(w, "Referer path must start with /cui/", http.StatusBadRequest)
-		return
-	}
-
-	parts := strings.SplitN(strings.TrimPrefix(refPath, "/cui/"), "/", 2)
-	if len(parts) == 0 || parts[0] == "" {
-		http.Error(w, "Could not determine cluster name from referer", http.StatusBadRequest)
-		return
-	}
-
-	clusterName := parts[0]
-	logger.Log.Info("API request to", zap.String("path", r.URL.Path), zap.String("cluster", clusterName), zap.String("referer", referer))
-
-	// Get the namespace from environment variable
-	namespace := os.Getenv("WATCH_NAMESPACE")
-	if namespace == "" {
-		http.Error(w, "WATCH_NAMESPACE environment variable not set", http.StatusInternalServerError)
-		return
-	}
-
-	// Create the service name for the Couchbase UI (following standard naming)
-	svcName := clusterName + "-ui"
-
-	// Create the target URL - using port 8091 which is the Couchbase UI port
-	targetURL := &url.URL{
-		Scheme: "http",
-		Host:   fmt.Sprintf("%s.%s.svc.cluster.local:8091", svcName, namespace),
-	}
-
-	logger.Log.Info("Proxying API request to", zap.String("targetURL", targetURL.String()))
-
-	// Create a reverse proxy
-	proxy := httputil.NewSingleHostReverseProxy(targetURL)
-
-	// Handle errors
-	proxy.ErrorHandler = func(w http.ResponseWriter, r *http.Request, err error) {
-		logger.Log.Error("API Proxy error", zap.Error(err))
-		http.Error(w, fmt.Sprintf("API Proxy error: %v", err), http.StatusBadGateway)
-	}
-
-	// Update the Host header and target
-	originalDirector := proxy.Director
-	proxy.Director = func(req *http.Request) {
-		// Set up the request
-		originalDirector(req)
-		req.Host = targetURL.Host
-		req.URL.Scheme = targetURL.Scheme
-		req.URL.Host = targetURL.Host
-		// Keep the original path
-		logger.Log.Info("Final API request", zap.String("requestURL", req.URL.String()))
-	}
-
-	// Serve the proxy request
-	proxy.ServeHTTP(w, r)
 }
 
 // this function is a goroutine that handles the websocket connection
@@ -702,97 +479,6 @@ func (s *Server) broadcastConditions() {
 			delete(s.clients, client)
 		}
 	}
-}
-
-// handleCouchbaseUIProxy handles reverse proxy requests to Couchbase UI
-func (s *Server) handleCouchbaseUIProxy(w http.ResponseWriter, r *http.Request) {
-	// Extract the cluster name from the URL path
-	// URL format: /cui/clustername/...
-	path := r.URL.Path
-	parts := strings.SplitN(strings.TrimPrefix(path, "/cui/"), "/", 2)
-
-	if len(parts) == 0 || parts[0] == "" {
-		http.Error(w, "Cluster name is required", http.StatusBadRequest)
-		return
-	}
-
-	clusterName := parts[0]
-	logger.Log.Info("Proxying to Couchbase UI for cluster", zap.String("cluster", clusterName), zap.String("path", path))
-
-	// Get the namespace from environment variable
-	namespace := os.Getenv("WATCH_NAMESPACE")
-	if namespace == "" {
-		http.Error(w, "WATCH_NAMESPACE environment variable not set", http.StatusInternalServerError)
-		return
-	}
-
-	// Create the service name for the Couchbase UI (following standard naming)
-	svcName := clusterName + "-ui"
-
-	// Create the target URL - using port 8091 which is the Couchbase UI port
-	targetURL := &url.URL{
-		Scheme: "http",
-		Host:   fmt.Sprintf("%s.%s.svc.cluster.local:8091", svcName, namespace),
-	}
-
-	//log.Printf("Proxying to: %s", targetURL.String())
-
-	// Create a reverse proxy
-	proxy := httputil.NewSingleHostReverseProxy(targetURL)
-
-	// Handle errors
-	proxy.ErrorHandler = func(w http.ResponseWriter, r *http.Request, err error) {
-		logger.Log.Error("Proxy error", zap.Error(err))
-		http.Error(w, fmt.Sprintf("Proxy error: %v", err), http.StatusBadGateway)
-	}
-
-	// Update the Host header and target
-	originalDirector := proxy.Director
-	proxy.Director = func(req *http.Request) {
-		// First, let the default director set up the request
-		originalDirector(req)
-
-		// Set the Host header
-		req.Host = targetURL.Host
-
-		// Create a new URL with the target scheme and host
-		req.URL.Scheme = targetURL.Scheme
-		req.URL.Host = targetURL.Host
-
-		// Strip /cui/clustername from the path
-		cuiPrefix := fmt.Sprintf("/cui/%s", clusterName)
-		strippedPath := strings.TrimPrefix(req.URL.Path, cuiPrefix)
-		req.URL.Path = strippedPath
-
-		//log.Printf("Proxying request: from %s to %s", path, req.URL.String())
-	}
-
-	// Modify the response to handle redirects and add base href
-	proxy.ModifyResponse = func(resp *http.Response) error {
-		// Handle redirects (3xx responses)
-		if resp.StatusCode >= 300 && resp.StatusCode < 400 {
-			location := resp.Header.Get("Location")
-			if location != "" {
-				logger.Log.Info("Original redirect location", zap.String("location", location))
-				redirectURL, err := url.Parse(location)
-				if err != nil {
-					logger.Log.Error("Error parsing redirect URL", zap.Error(err))
-				} else {
-					// Extract the path and rewrite it to our proxy format
-					path := redirectURL.Path
-					newLocation := fmt.Sprintf("/cui/%s%s", clusterName, path)
-					//log.Printf("Rewriting absolute redirect to: %s", newLocation)
-					resp.Header.Set("Location", newLocation)
-				}
-
-			}
-		}
-
-		return nil
-	}
-
-	// Serve the proxy request directly without path rewriting
-	proxy.ServeHTTP(w, r)
 }
 
 // sendCachedEvents sends cached events for a cluster to a specific client
